refactor: 쿠폰 서비스 디렉토리 위치 변경 - DDD에 부합하지 않는 위치 수정
